pyproject.toml
[tool.poetry]
name = "dvmn-support-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dvmn_support_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.4"
python-dotenv = "^1.0.0"
google-cloud-dialogflow = "^2.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
